chore: properly install an entrypoint script for the kernel_crawler.cli function.
