+++ docker/Dockerfile
@@ -3,7 +3,7 @@ FROM python:3.8-slim-buster
 WORKDIR /app
 
 COPY . .
-RUN pip3 install -e .
+RUN pip3 install .
 
 COPY ./docker/root /
