+++ docker/root/usr/bin/entrypoint
@@ -1,2 +1,3 @@
 #!/bin/bash
-python3 /app/__init__.py crawl "$@"
+
+kernel-crawler "$@"
